server: move run_raft_server into tikv
